SCRUM-42 feature\add landing page
SCRUM-42 add a landing and homepage for users before and after login
